Create W1-05
Declare a struct called RationalNumber that has two integer member variables. Call the member variables
a (to mean the numerator) and b (to mean the denominator). Then write a function named toDouble that
takes one RationalNumber argument and returns the rational number as double. Write a main program to
test your function.