Create W1-07
Write a function named reduce that takes one RationalNumber argument by reference and reduces the
rational number. Reduction of rational number means to write the rational number in the simplest form
possible. Example 7/21 is reduced to 1/3; 2/4 is reduced to 1/2; 15/5 is reduced to 3/1 and so on so forth.
Don't jump to write program. Instead think carefully how do we do reduction by hand. Hint:- What is the
definition of GCD? What does it have to do with reduction of rational numbers? What does the function
return? Write a main program to test your function